+++ substrate/frame/revive/rpc/src/subxt_client.rs
@@ -64,7 +64,7 @@ pub use subxt::config::PolkadotConfig as SrcChainConfig;
 		with = "::subxt::utils::Static<::sp_weights::Weight>"
 	),
 	substitute_type(
-		path = "pallet_revive::evm::api::rpc_types_gen::Block",
+		path = "pallet_revive::evm::api::block::Block",
 		with = "::subxt::utils::Static<::pallet_revive::evm::Block>"
 	),
 	substitute_type(
